Add files via upload
Parsing value from a string.
In many cases we need to pass the value from a string. It is meaning that we have to convert the string into some other data type such as a number string into an int. We might need to convert a number string. It is when reading input from the user perhaps after it's been entered into a console or after being retrieved from a user interface.